service/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
langchain
langchain-community
langchain-core
langchain-groq
langchain-text-splitters
langchain_huggingface
sentence-transformers
duckduckgo-search
wikipedia
arxiv
pillow
pytesseract
googlemaps